requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs==25.3.0
augmax==0.4.1
av==14.4.0
beartype==0.19.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
chex==0.1.89
click==8.2.1
cloudpickle==3.1.1
cmake==4.0.2
comm==0.2.2
contourpy==1.3.2
crc32c==2.7.1
cycler==0.12.1
datasets==4.0.0
debugpy==1.8.14
decorator==5.2.1
decord==0.6.0
deepdiff==8.5.0
diffusers==0.33.1
dill==0.3.8
distlib==0.3.9
dm-control==1.0.14
dm-env==1.6
dm-tree==0.1.9
docker-pycreds==0.4.0
docstring_parser==0.16
donfig==0.8.1.post1
draccus==0.10.0
einops==0.8.1
equinox==0.12.2
etils==1.12.2
evdev==1.9.2
executing==2.2.0
Farama-Notifications==0.0.4
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
flax==0.10.2
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2025.3.0
gcsfs==2025.3.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
glfw==2.9.0
google-api-core==2.24.2
google-auth==2.40.2
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gym-aloha==0.1.1
gymnasium==0.29.1
h5py==3.13.0
hf-xet==1.1.2
hf_transfer==0.1.9
huggingface-hub==0.32.3
humanize==4.12.3
identify==2.6.12
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
inquirerpy==0.3.4
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jax==0.5.3
jax-cuda12-pjrt==0.5.3
jax-cuda12-plugin==0.5.3
jaxlib==0.5.3
jaxtyping==0.2.36
jedi==0.19.2
Jinja2==3.1.6
jsonlines==4.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
labmaze==1.0.6
lerobot @ git+https://github.com/huggingface/lerobot@0cf864870cf29f4738d3ade893e6fd13fbd7cdb5
llvmlite==0.44.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
ml-dtypes==0.4.1
ml_collections==1.0.0
mpmath==1.3.0
msgpack==1.1.0
mujoco==2.3.7
multidict==6.4.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nodeenv==1.9.1
numba==0.61.2
numcodecs==0.16.1
numpy==1.26.4
numpydantic==1.6.9
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvcc-cu12==12.9.41
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
-e git+https://github.com/Physical-Intelligence/openpi.git@b29f81b6376f84606278a3350732a6cc806bb5e8#egg=openpi
-e git+https://github.com/Physical-Intelligence/openpi.git@b29f81b6376f84606278a3350732a6cc806bb5e8#egg=openpi_client&subdirectory=packages/openpi-client
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.11.13
orderly-set==5.4.1
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pfzy==0.3.4
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
polars==1.30.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pymunk==7.0.0
pynput==1.8.1
pynvml==12.0.0
PyOpenGL==3.1.9
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-xlib==0.33
pytz==2025.2
PyYAML==6.0.2
pyyaml-include==1.4.1
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rerun-sdk==0.23.1
rich==14.0.0
rsa==4.9.1
ruff==0.11.12
safetensors==0.5.3
scipy==1.15.3
sentencepiece==0.2.0
sentry-sdk==2.29.1
setproctitle==1.3.6
shtab==1.7.2
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
soupsieve==2.7
stack-data==0.6.3
svgwrite==1.4.3
sympy==1.14.0
tensorstore==0.1.74
termcolor==3.1.0
tokenizers==0.21.1
toml==0.10.2
toolz==1.0.0
torch==2.7.1
torchcodec==0.4.0
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
tqdm-loggable==0.2
traitlets==5.14.3
transformers==4.53.2
Tree==0.2.4
treescope==0.1.9
triton==3.3.1
typeguard==4.4.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tyro==0.9.22
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.31.2
wadler_lindig==0.1.6
wandb==0.19.11
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.14.1
xxhash==3.5.0
yarl==1.20.0
zarr==3.0.8
zipp==3.22.0
